added custom and subdomain routing to page.tsx also removed firebase hosting
